Score calculation
Function to calculate each teams score. Returning the object that was suggested in the issue description.
Not sure how to try this function on the game server - or where to put the function call.
But I put together a simulation on my laptop with the game state-example and it works as expected.

Closes #10