Airbnb (and other) error fixing
Since npm run check didn't work for me (using Windows) I tried using npx eslint . --quiet instead and I got a whole boatload of errors, most were of the type "Expected linebreaks to be 'LF' but found 'CRLF'  linebreak-style". There were some "legitamate" errors in there as well but the linebreak-errors were so many it was hard to weed out the other ones. To get rid of all the linebreak-errors i tried using npx eslint . --fix to get rid of them and it seems like it fixed some of the "legitimate" errors as well. The last errors were mostly that I had completely missed that we sent the context to a spell or event as a argument, so I added in the context as a parameter where they were missing. I also fixed the last ones which weren't that kind of problem.