Added visual infomation
Added visual information in the form of the opacity getting lower when the secret cargo isn't active by subscribing to the secret-cargo:faded event and switching classes when the event is recieved. Also playing around a bit with the timing of when the timer runs out/when the cargo is set as inactive.